requirements/development.txt
-r base.txt

# Debug
django-debug-toolbar>=4.4
django-browser-reload>=1.0

# Testing
pytest>=8.0
pytest-django>=4.9
pytest-cov>=5.0
factory-boy>=3.3
faker>=30.0

# Linting
ruff>=0.8
mypy>=1.13
django-stubs>=5.1
pre-commit>=4.0
